- hotfix case insensitivity for categories - implementation of subcategory for wealth and query for view of the wealth and liquidity comparison
