+++ .gitignore
@@ -1,1 +1,1 @@
-/webapp/tracker.db
+webapp/tracker.db
