+++ webapp/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "1.3.0"
+__version__ = "1.4.1"
