--- webapp/static/style.css
@@ -1,7 +1,0 @@
-.container {
-  display: flex;
-}
-
-.left, .categories {
-  flex: 1;
-}
