+++ webapp/templates/categories.html
@@ -5,7 +5,10 @@
         <form method="POST">
             <h3>Add personal category</h3>
                 <div class="form-group">
-                    <input type="text" class="form-group" id="category_add" name="category_add" placeholder="Your custom category">
+                    <input type="text" class="form-group" id="category_add" name="category_add" placeholder="Your custom category"><br>
+
+                    <label style="text-align:left; .form-group {width: 40px; height: 40px; ">Check if the category is related to increase of wealth?  <input type="checkbox" class="form-group" id="is-wealth" name="is-wealth"> </label>
+
                 </div>
             <button type="submit" name="submit_category" class="btn btn-primary">add category</button>
         </form>
